adjust how to stay up to date on policy changes
The privacy policy landscape is changing rapidly, which could result in lots of emails or in-app notifications in a short period. That could feel very disruptive to customers so we've adjusted how to to keep up with changes in our policies. Both our blog and the Github repository offer ways to subscribe for email notifications.