Introduce Rollbar as a HEY subprocessor for error reporting
DPA signed 2020-08-31.

We use Rollbar to track & report on internal system errors in HEY.

Error data reported to Rollbar includes information about the web
request that triggered the error and about the internal system state
at the time of the error.

We scrub sensitive and personal data before sending to Rollbar.
However, we may miss some source of personal data, so we treat Rollbar
as a subprocessor and signed a DPA with them.