update to reflect HEY's tax-exempt status in the US
The US has a federal law called the Internet Tax Freedom Act. Under the ITFA, Internet access is not subject to sales tax. After several months of seeking counsel from tax authorities, we got confirmation the ITFA applies to HEY because as an email service, it falls under the ITFA's definition of Internet access. After getting this confirmation, we stopped collecting and remitting sales tax on HEY for US-based customers and refunded all HEY customers in the US who had paid sales tax for HEY to date.