be even more informative about GDPR
The Schrems II ruling has opened up a lot of questions. In this update, we try to answer as many of the questions we can upfront about the need for the DPA and the due diligence we've taken as a company to uphold our responsibilities under the Standard Contractual Clauses. We are also preparing for the roll-in of our data processing addendum into the Terms of Service, in lieu of the current opt-in only approach. The updated DPA, which incorporates both controller-processor and controller-controller SCCs, will be added to our policy repo in a separate commit.